CNV-32374: Move template params to Catalog drawer
Move setting of the template parameters that used to be present in
"Customize VirtualMachine" page one step back - to the Catalog drawer.
Remove "Customize VirtualMachine" page.

Fixes https://issues.redhat.com/browse/CNV-32374
