+++ locales/en/plugin__kubevirt-plugin.json
@@ -312,7 +312,6 @@
   "Customize and create VirtualMachine": "Customize and create VirtualMachine",
   "Customize template parameters": "Customize template parameters",
   "Customize VirtualMachine": "Customize VirtualMachine",
-  "Customize VirtualMachine parameters": "Customize VirtualMachine parameters",
   "CX series": "CX series",
   "CX Series": "CX Series",
   "cx1": "cx1",
@@ -1069,7 +1068,7 @@
   "This is a CD-ROM boot source": "This is a CD-ROM boot source",
   "This key will override the SSH key secret set on the template": "This key will override the SSH key secret set on the template",
   "This Persistent Volume Claim will be created using a DataVolume through Containerized Data Importer (CDI)": "This Persistent Volume Claim will be created using a DataVolume through Containerized Data Importer (CDI)",
-  "This Template requires some additional parameters. Click the Customize VirtualMachine button to complete the creation flow.": "This Template requires some additional parameters. Click the Customize VirtualMachine button to complete the creation flow.",
+  "This Template requires some additional parameters.": "This Template requires some additional parameters.",
   "This Template supports quick create VirtualMachine": "This Template supports quick create VirtualMachine",
   "This user is not allowed to edit this boot source": "This user is not allowed to edit this boot source",
   "This VirtualMachine has": "This VirtualMachine has",
