--- src/views/catalog/customize/CustomizeVirtualMachine.scss
@@ -1,19 +1,0 @@
-.customize-vm {
-  &__right-header {
-    a {
-      padding-left: 0;
-    }
-
-    img {
-      width: 2em;
-    }
-  }
-
-  .pf-c-expandable-section__toggle {
-    align-items: center;
-
-    &-text {
-      font-size: 18px;
-    }
-  }
-}
