+++ src/views/catalog/customize/components/CustomizeForms/CustomizeForm.tsx
@@ -1,15 +1,12 @@
-import React, { FC, useMemo } from 'react';
+import React, { FC } from 'react';
 import { FormProvider, useForm } from 'react-hook-form';
 
-import { getTemplateNameParameterValue } from '@catalog/customize/utils';
 import { V1Template } from '@kubevirt-ui/kubevirt-api/console';
-import { useURLParams } from '@kubevirt-utils/hooks/useURLParams';
 import { Form } from '@patternfly/react-core';
 
-import { buildFields, getVirtualMachineNameField } from '../../utils';
+import { buildFields } from '../../utils';
 import { ExpandableOptionsFields } from '../ExpandableOptionalFields';
 import { FieldGroup } from '../FieldGroup';
-import { FormActionGroup } from '../FormActionGroup';
 import { FormError } from '../FormError';
 
 import { useCustomizeFormSubmit } from './useCustomizeFormSubmit';
@@ -22,18 +19,12 @@ type CustomizeFormProps = {
 export const CustomizeForm: FC<CustomizeFormProps> = ({ template }) => {
   const methods = useForm();
 
-  const { error, loaded, onSubmit } = useCustomizeFormSubmit({ template });
+  const { error, onSubmit } = useCustomizeFormSubmit({ template });
   const [requiredFields, optionalFields] = buildFields(template);
-
-  const { params } = useURLParams();
-  const vmName = params.get('vmName') || getTemplateNameParameterValue(template);
-
-  const nameField = useMemo(() => getVirtualMachineNameField(vmName), [vmName]);
 
   return (
     <FormProvider {...methods}>
       <Form onSubmit={methods.handleSubmit(onSubmit)}>
-        <FieldGroup field={nameField} showError={error} />
         {requiredFields?.map((field) => (
           <FieldGroup field={field} key={field.name} showError={error} />
         ))}
@@ -41,7 +32,6 @@ export const CustomizeForm: FC<CustomizeFormProps> = ({ template }) => {
         <ExpandableOptionsFields optionalFields={optionalFields} />
 
         <FormError error={error} />
-        <FormActionGroup loading={!loaded} />
       </Form>
     </FormProvider>
   );
