+++ src/views/catalog/customize/components/CustomizeForms/CustomizeFormWithStorage.tsx
@@ -1,10 +1,8 @@
 import React, { FC, useMemo, useState } from 'react';
 import { FormProvider, useForm } from 'react-hook-form';
 
-import { getTemplateNameParameterValue } from '@catalog/customize/utils';
 import { V1Template } from '@kubevirt-ui/kubevirt-api/console';
 import { V1beta1DataVolumeSpec, V1ContainerDiskSource } from '@kubevirt-ui/kubevirt-api/kubevirt';
-import { useURLParams } from '@kubevirt-utils/hooks/useURLParams';
 import {
   getTemplateOS,
   getTemplateVirtualMachineObject,
@@ -14,15 +12,10 @@ import {
 import { Form } from '@patternfly/react-core';
 
 import { addCDToTemplate } from '../../cd';
-import {
-  buildFields,
-  getVirtualMachineNameField,
-  overrideVirtualMachineDataVolumeSpec,
-} from '../../utils';
+import { buildFields, overrideVirtualMachineDataVolumeSpec } from '../../utils';
 import { ExpandableCustomizeSourceSection } from '../CustomizeSource/ExpandableCustomizeSourceSection';
 import { ExpandableOptionsFields } from '../ExpandableOptionalFields';
 import { FieldGroup } from '../FieldGroup';
-import { FormActionGroup } from '../FormActionGroup';
 import { FormError } from '../FormError';
 
 import { useCustomizeFormSubmit } from './useCustomizeFormSubmit';
@@ -37,8 +30,6 @@ const CustomizeFormWithStorage: FC<CustomizeFormWithStorageProps> = ({
   template,
 }) => {
   const methods = useForm();
-  const { params } = useURLParams();
-  const vmName = params.get('vmName') || getTemplateNameParameterValue(template);
 
   const [diskSource, setDiskSource] = useState<V1beta1DataVolumeSpec>();
   const [cdSource, setCDSource] = useState<V1beta1DataVolumeSpec | V1ContainerDiskSource>();
@@ -62,7 +53,7 @@ const CustomizeFormWithStorage: FC<CustomizeFormWithStorageProps> = ({
     getTemplateOS(template) === OS_NAME_TYPES.windows,
   );
 
-  const { cdUpload, diskUpload, error, loaded, onCancel, onSubmit } = useCustomizeFormSubmit({
+  const { cdUpload, diskUpload, error, onSubmit } = useCustomizeFormSubmit({
     cdSource,
     diskSource,
     template: templateWithSources,
@@ -71,13 +62,9 @@ const CustomizeFormWithStorage: FC<CustomizeFormWithStorageProps> = ({
 
   const [requiredFields, optionalFields] = useMemo(() => buildFields(template), [template]);
 
-  const nameField = useMemo(() => getVirtualMachineNameField(vmName), [vmName]);
-
   return (
     <FormProvider {...methods}>
       <Form onSubmit={methods.handleSubmit(onSubmit)}>
-        <FieldGroup field={nameField} showError={error} />
-
         {requiredFields?.map((field) => (
           <FieldGroup field={field} key={field.name} showError={error} />
         ))}
@@ -98,7 +85,6 @@ const CustomizeFormWithStorage: FC<CustomizeFormWithStorageProps> = ({
         <ExpandableOptionsFields optionalFields={optionalFields} />
 
         <FormError error={error} />
-        <FormActionGroup loading={!loaded} onCancel={onCancel} />
       </Form>
     </FormProvider>
   );
