+++ src/views/catalog/templatescatalog/components/TemplatesCatalogDrawer/TemplatesCatalogDrawerPanel.tsx
@@ -1,54 +1,31 @@
-import React, { FC, memo, useEffect, useState } from 'react';
+import React, { FC, memo, useEffect, useMemo, useState } from 'react';
 import { useParams } from 'react-router-dom';
 import produce from 'immer';
 
+import { CustomizeForm } from '@catalog/customize/components/CustomizeForms/CustomizeForm';
+import CustomizeFormWithStorage from '@catalog/customize/components/CustomizeForms/CustomizeFormWithStorage';
+import { hasCustomizableSource } from '@catalog/customize/utils';
 import { updateVMCPUMemory } from '@catalog/templatescatalog/utils/helpers';
 import { useWizardVMContext } from '@catalog/utils/WizardVMContext';
-import { WizardOverviewDisksTable } from '@catalog/wizard/tabs/overview/components/WizardOverviewDisksTable/WizardOverviewDisksTable';
-import { WizardOverviewNetworksTable } from '@catalog/wizard/tabs/overview/components/WizardOverviewNetworksTable/WizardOverviewNetworksTable';
 import { ProcessedTemplatesModel, V1Template } from '@kubevirt-ui/kubevirt-api/console';
 import { V1VirtualMachine } from '@kubevirt-ui/kubevirt-api/kubevirt';
-import AdditionalResources from '@kubevirt-utils/components/AdditionalResources/AdditionalResources';
-import CPUDescription from '@kubevirt-utils/components/CPUDescription/CPUDescription';
-import { CpuMemHelperTextResources } from '@kubevirt-utils/components/CPUDescription/utils/utils';
-import CPUMemory from '@kubevirt-utils/components/CPUMemory/CPUMemory';
-import CPUMemoryModal from '@kubevirt-utils/components/CPUMemoryModal/CpuMemoryModal';
-import HardwareDevices from '@kubevirt-utils/components/HardwareDevices/HardwareDevices';
-import { useModal } from '@kubevirt-utils/components/ModalProvider/ModalProvider';
-import VirtualMachineDescriptionItem from '@kubevirt-utils/components/VirtualMachineDescriptionItem/VirtualMachineDescriptionItem';
 import { DEFAULT_NAMESPACE } from '@kubevirt-utils/constants/constants';
 import { useKubevirtTranslation } from '@kubevirt-utils/hooks/useKubevirtTranslation';
-import { WORKLOADS_LABELS } from '@kubevirt-utils/resources/template/utils/constants';
-import {
-  getTemplateDescription,
-  getTemplateDisks,
-  getTemplateDocumentationURL,
-  getTemplateInterfaces,
-  getTemplateName,
-  getTemplateNetworks,
-  getTemplateVirtualMachineObject,
-  getTemplateWorkload,
-  isDefaultVariantTemplate,
-} from '@kubevirt-utils/resources/template/utils/selectors';
-import { getCPU, getGPUDevices, getHostDevices } from '@kubevirt-utils/resources/vm';
+import { useURLParams } from '@kubevirt-utils/hooks/useURLParams';
+import { getTemplateVirtualMachineObject } from '@kubevirt-utils/resources/template/utils/selectors';
 import { k8sCreate } from '@openshift-console/dynamic-plugin-sdk';
 import {
   Alert,
   AlertVariant,
-  Button,
-  DescriptionList,
-  DescriptionListDescription,
-  DescriptionListGroup,
-  DescriptionListTerm,
   Grid,
   GridItem,
   Stack,
   StackItem,
   Title,
 } from '@patternfly/react-core';
-import ExternalLinkSquareAltIcon from '@patternfly/react-icons/dist/esm/icons/external-link-square-alt-icon';
 
-import TemplateExpandableDescription from './TemplateExpandableDescription';
+import TemplatesCatalogDrawerLeftColumn from './TemplatesCatalogDrawerLeftColumn';
+import TemplatesCatalogDrawerRightColumn from './TemplatesCatalogDrawerRightColumn';
 
 type TemplatesCatalogDrawerPanelProps = {
   template: V1Template;
@@ -57,27 +34,27 @@ type TemplatesCatalogDrawerPanelProps = {
 export const TemplatesCatalogDrawerPanel: FC<TemplatesCatalogDrawerPanelProps> = memo(
   ({ template }) => {
     const { t } = useKubevirtTranslation();
-    const { createModal } = useModal();
     const { updateVM } = useWizardVMContext();
     const { ns } = useParams<{ ns: string }>();
-    const vmNamespace = ns || DEFAULT_NAMESPACE;
+    const { params } = useURLParams();
 
-    const notAvailable = t('N/A');
-    const vmObject = getTemplateVirtualMachineObject(template);
-    const displayName = getTemplateName(template);
-    const description = getTemplateDescription(template) || notAvailable;
-    const documentationUrl = getTemplateDocumentationURL(template);
-    const workload = getTemplateWorkload(template);
-    const networks = getTemplateNetworks(template);
-    const interfaces = getTemplateInterfaces(template);
-    const disks = getTemplateDisks(template);
-    const isDefaultTemplate = isDefaultVariantTemplate(template);
-    const hostDevicesCount = getHostDevices(vmObject)?.length || 0;
-    const gpusCount = getGPUDevices(vmObject)?.length || 0;
-    const hardwareDevicesCount = hostDevicesCount + gpusCount;
+    const isBootSourceAvailable = params.get('defaultSourceExists') === 'true';
+    const vmNamespace = ns || DEFAULT_NAMESPACE;
 
     const [updatedVM, setUpdatedVM] = useState<V1VirtualMachine>(undefined);
     const [error, setError] = useState(undefined);
+
+    const [templateWithGeneratedValues, setTemplateWithGeneratedValues] = useState(template);
+
+    const Form = useMemo(() => {
+      const withDiskSource = hasCustomizableSource(template);
+
+      if (withDiskSource) {
+        return CustomizeFormWithStorage;
+      } else {
+        return CustomizeForm;
+      }
+    }, [template]);
 
     useEffect(() => {
       setError(undefined);
@@ -94,6 +71,7 @@ export const TemplatesCatalogDrawerPanel: FC<TemplatesCatalogDrawerPanelProps> =
         },
       })
         .then((processedTemplate) => {
+          setTemplateWithGeneratedValues(processedTemplate);
           updateVMCPUMemory(
             vmNamespace,
             updateVM,
@@ -121,106 +99,25 @@ export const TemplatesCatalogDrawerPanel: FC<TemplatesCatalogDrawerPanelProps> =
               <StackItem>
                 <Grid hasGutter>
                   <GridItem span={6}>
-                    <DescriptionList>
-                      <DescriptionListGroup>
-                        <DescriptionListTerm>{t('Operating system')}</DescriptionListTerm>
-                        <DescriptionListDescription>{displayName}</DescriptionListDescription>
-                      </DescriptionListGroup>
-                      <DescriptionListGroup>
-                        <DescriptionListTerm>{t('Workload type')}</DescriptionListTerm>
-                        <DescriptionListDescription>
-                          {WORKLOADS_LABELS[workload] ?? t('Other')}{' '}
-                          {isDefaultTemplate && t('(default)')}
-                        </DescriptionListDescription>
-                      </DescriptionListGroup>
-                      <DescriptionListGroup>
-                        <DescriptionListTerm>{t('Description')}</DescriptionListTerm>
-                        <DescriptionListDescription>
-                          {<TemplateExpandableDescription description={description} />}
-                        </DescriptionListDescription>
-                      </DescriptionListGroup>
-                      <DescriptionListGroup>
-                        <DescriptionListTerm>{t('Documentation')}</DescriptionListTerm>
-                        <DescriptionListDescription>
-                          {documentationUrl ? (
-                            <Button
-                              icon={<ExternalLinkSquareAltIcon />}
-                              iconPosition="right"
-                              isInline
-                              isSmall
-                              variant="link"
-                            >
-                              <a href={documentationUrl} rel="noopener noreferrer" target="_blank">
-                                {t('Refer to documentation')}
-                              </a>
-                            </Button>
-                          ) : (
-                            notAvailable
-                          )}
-                        </DescriptionListDescription>
-                      </DescriptionListGroup>
-                      <AdditionalResources template={template} />
-                    </DescriptionList>
+                    <TemplatesCatalogDrawerLeftColumn template={template} />
                   </GridItem>
                   <GridItem span={6}>
-                    <DescriptionList>
-                      <VirtualMachineDescriptionItem
-                        bodyContent={
-                          <CPUDescription
-                            cpu={getCPU(updatedVM)}
-                            helperTextResource={CpuMemHelperTextResources.FutureVM}
-                          />
-                        }
-                        onEditClick={() =>
-                          createModal(({ isOpen, onClose }) => (
-                            <CPUMemoryModal
-                              isOpen={isOpen}
-                              onClose={onClose}
-                              onSubmit={updateVMCPUMemory(vmNamespace, updateVM, setUpdatedVM)}
-                              templateNamespace={template?.metadata?.namespace}
-                              vm={updatedVM}
-                            />
-                          ))
-                        }
-                        descriptionData={<CPUMemory vm={updatedVM} />}
-                        descriptionHeader={t('CPU | Memory')}
-                        isEdit
-                        isPopover
-                      />
-                      <DescriptionListGroup>
-                        <DescriptionListTerm>
-                          {t('Network interfaces')}
-                          {` (${networks.length})`}
-                        </DescriptionListTerm>
-                        <DescriptionListDescription>
-                          <WizardOverviewNetworksTable
-                            interfaces={interfaces}
-                            networks={networks}
-                          />
-                        </DescriptionListDescription>
-                      </DescriptionListGroup>
-                      <DescriptionListGroup>
-                        <DescriptionListTerm>
-                          {t('Disks')}
-                          {` (${disks.length})`}
-                        </DescriptionListTerm>
-                        <DescriptionListDescription>
-                          <WizardOverviewDisksTable vm={vmObject} />
-                        </DescriptionListDescription>
-                      </DescriptionListGroup>
-                      <DescriptionListGroup>
-                        <DescriptionListTerm>
-                          {t('Hardware devices')}
-                          {` (${hardwareDevicesCount})`}
-                        </DescriptionListTerm>
-                        <DescriptionListDescription>
-                          <HardwareDevices hideEdit vm={vmObject} />
-                        </DescriptionListDescription>
-                      </DescriptionListGroup>
-                    </DescriptionList>
+                    <TemplatesCatalogDrawerRightColumn
+                      setUpdatedVM={setUpdatedVM}
+                      template={template}
+                      updatedVM={updatedVM}
+                    />
                   </GridItem>
                 </Grid>
               </StackItem>
+
+              <StackItem>
+                <Form
+                  isBootSourceAvailable={isBootSourceAvailable}
+                  template={templateWithGeneratedValues}
+                />
+              </StackItem>
+
               {error && (
                 <StackItem>
                   <Alert isInline title={t('Error')} variant={AlertVariant.danger}>
